# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Sonit Server",
        "title": "Sonit Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/collections": {
            "get": {
                "description": "Retrieve all collections from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get all collections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Collection"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/activate/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activate a collection by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Activate a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new collection with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Create a new collection",
                "parameters": [
                    {
                        "description": "Collection creation request",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/name/{name}": {
            "get": {
                "description": "Retrieve collections that match the specified name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collections by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Collection"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/remove/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove/delete a collection by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Remove a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve collections filtered by their status (active/inactive)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collections by status",
                "parameters": [
                    {
                        "description": "Collection status (true for active, false for inactive)",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Collection"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing collection with new details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Update a collection",
                "parameters": [
                    {
                        "description": "Collection update request",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/collections/{id}": {
            "get": {
                "description": "Retrieve a specific collection by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collection by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Collection"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Collection": {
            "type": "object",
            "properties": {
                "active_status": {
                    "type": "boolean"
                },
                "collection_id": {
                    "type": "string"
                },
                "collection_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.CreateCollectionRequest": {
            "type": "object",
            "required": [
                "collection_name"
            ],
            "properties": {
                "collection_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateCollectionRequest": {
            "type": "object",
            "required": [
                "collection_id"
            ],
            "properties": {
                "collection_id": {
                    "type": "string"
                },
                "collection_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
        "response.MessageAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}